+++ narco/conf.py
@@ -1,0 +1,5 @@
+import os
+
+cartel_dir = os.path.join(os.path.expanduser("~"), ".cartel")
+
+CARTEL_URL = "http://127.0.0.1:5000"
